requirements.txt
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.3
defusedxml==0.7.1
Django==6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-spectacular==0.29.0
drf-standardized-errors==0.15.0
drf-yasg==1.21.14
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
icalendar==6.3.2
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
oauthlib==3.3.1
packaging==26.0
paystackapi==2.1.3
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.10.1
pypaystack2==2.0.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.30.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.7.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
gunicorn==23.0.0
whitenoise
dj-database-url==2.3.0
